use Formats.alwaysEscapeUnicode at Jackson
fix https://github.com/json4s/json4s/issues/750

In Jackson, the render method did not reference the value of Formats.alwaysEscapeUnicode.
Therefore, it was not in line with the native behavior.

Fix this problem.

Also, Jackson always uses uppercase when escaping UNICODE.
To make the results consistent, the native module now also outputs in uppercase.